Re arranged the structure files (#8)
