+++ src/components/Form.css
@@ -1,7 +1,24 @@
 .form-container {
   padding: 50px 0;
   text-align: center;
+  display: flex;
+  justify-content: center;
+  position: relative;
+  z-index: 20;
 }
+
+.star {
+  position: absolute;
+  top: -100px
+}
+
+.form-wrapper {
+  background-color: #00000052;
+  border-radius: 10px;
+  max-width: 700px;
+  padding: 50px 20px
+}
+
 
 /* Styling for questions */
 .question-container {
@@ -9,6 +26,7 @@
   padding: 0 30px;
   text-align: center;
   margin: 80px auto;
+  margin-top: 20px;
 }
 
 .count {
