--- src/components/SubmitButton.jsx
@@ -1,10 +1,0 @@
-import "./Form.css";
-
-
-export const SubmitButton = () => {
-  return (
-    <div>
-    <button>Submit your answer</button>
-    </div>
-  )
-}
