+++ src/components/Summary.jsx
@@ -1,0 +1,10 @@
+
+
+export const Summary = ({ value, mood, song }) => {
+  return (
+    <div>
+      <h2>Summary of The Science of Happiness</h2>
+      <p>{value} is your mood-boosting activity. {mood} music is making you happy, and if ever you are down - listen to {song}</p>
+      </div>
+  )
+}
