+++ src/index.css
@@ -15,9 +15,9 @@
   -moz-osx-font-smoothing: grayscale;
 
   /* CSS - Variables */
-  --primary: #0018a4;
-  --white: #fff;
-  --button: #c2c2c2;
+  --primary: #ffffff;
+  --white: #ffff;
+  --button: #6f6f6f;
   --radius: 6px;
 }
 
@@ -30,6 +30,20 @@ body {
   font-family: "Poppins", sans-serif;
   max-width: 100vw;
   min-height: 100vh;
+  background-image: url("src/assets/chuttersnap-hJNNvvHo3zw-unsplash.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+  position: relative;
+}
+
+
+body::before {
+  content: "";
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  background-image: linear-gradient(to bottom, rgba(0, 0, 0, 0.44), rgba(4, 50, 95, 0.44));
+
 }
 
 main,
@@ -60,10 +74,10 @@ h3 {
 
 /* Elements */
 button {
-  background: var(--button);
+  background: #d48247;
   color: var(--white);
   font-size: 20px;
-  border: 1px solid #000;
+  border: 1px solid var(--white);
   border-radius: var(--radius);
   padding: 18px 20px;
 }
